// packages/admin/src/enums/role.ts
export enum RolePermission {
  ADMIN_DASHBOARD = 'admin.dashboard',
  ADMIN_BANNER = 'admin.banner',
  ADMIN_PARTNER = 'admin.partner',
  ADMIN_CONTACT = 'admin.contact',
  ADMIN_TICKET = 'admin.ticket',
  ADMIN_NEWS = 'admin.news',
  ADMIN_MANAGE_ADMIN = 'admin.manage_admin',
  ADMIN_DEDICATED = 'admin.dedicated',
  ADMIN_VPS_AUTO = 'admin.vps_auto',
  ADMIN_VPS_SERVER = 'admin.vps_server',
  ADMIN_COLOCATION = 'admin.colocation',
  ADMIN_HOSTING = 'admin.hosting',
  ADMIN_DOMAIN = 'admin.domain',
  ADMIN_LICENSE = 'admin.license',
  ADMIN_ADDON = 'admin.addon',
  ADMIN_MANAGE_USER = 'admin.manage_user',
  ADMIN_ORDER = 'admin.order',
  ADMIN_PAYMENT = 'admin.payment',
  ADMIN_QUOTATION = 'admin.quotation',
  ADMIN_INVOICE = 'admin.invoice',
  ADMIN_RECEIPT = 'admin.receipt',
  ADMIN_BILLING_SLIP = 'admin.billing_slip',
  ADMIN_TEMPLATE = 'admin.template',
  ADMIN_MANAGE_IP = 'admin.manage_ip',
  ADMIN_DATA_CENTER = 'admin.data_center',
  ADMIN_RACK = 'admin.rack',
  ADMIN_SERVER = 'admin.server',
  ADMIN_MANAGE_MEMBERSHIP = 'admin.manage_membership',
  ADMIN_DOCUMENT = 'admin.document',
}